refactor: extract roles, notifications, wait, feedback, managers into tools/
Second batch of 5 modules extracted (21 tools moved). Total tool count
unchanged at 52. Remaining 7 modules follow in Task 8.
