+++ iterm_mcpy/fastmcp_server.py
@@ -2798,807 +2798,23 @@ Coordinate responses as needed using write_to_sessions or send_cascade_message.
 
 # ============================================================================
 # NOTIFICATION TOOLS
+# See iterm_mcpy/tools/notifications.py
 # ============================================================================
-
-@mcp.tool()
-async def get_notifications(request: GetNotificationsRequest, ctx: Context) -> str:
-    """Get recent agent notifications.
-
-    Returns a list of notifications about agent status changes, errors,
-    completions, and other events. Use this to stay aware of what's happening
-    across all managed agents.
-    """
-    notification_manager = ctx.request_context.lifespan_context["notification_manager"]
-    logger = ctx.request_context.lifespan_context["logger"]
-
-    try:
-        req = ensure_model(GetNotificationsRequest, request)
-        notifications = await notification_manager.get(
-            limit=req.limit,
-            level=req.level,
-            agent=req.agent,
-            since=req.since,
-        )
-
-        response = GetNotificationsResponse(
-            notifications=notifications,
-            total_count=len(notifications),
-            has_more=len(notifications) == req.limit,
-        )
-
-        logger.info(f"Retrieved {len(notifications)} notifications")
-        return response.model_dump_json(indent=2, exclude_none=True)
-
-    except Exception as e:
-        logger.error(f"Error getting notifications: {e}")
-        return f"Error: {e}"
-
-
-@mcp.tool()
-async def get_agent_status_summary(ctx: Context) -> str:
-    """Get a compact status summary of all agents.
-
-    Returns a one-line-per-agent summary showing the most recent
-    notification for each agent, including lock counts. Great for quick status checks.
-    """
-    notification_manager = ctx.request_context.lifespan_context["notification_manager"]
-    agent_registry = ctx.request_context.lifespan_context["agent_registry"]
-    lock_manager = ctx.request_context.lifespan_context.get("tag_lock_manager")
-    logger = ctx.request_context.lifespan_context["logger"]
-
-    try:
-        # Get latest notification per agent
-        latest = await notification_manager.get_latest_per_agent()
-
-        # Also include agents with no notifications
-        all_agents = agent_registry.list_agents()
-        for agent in all_agents:
-            if agent.name not in latest:
-                # Create a placeholder notification
-                latest[agent.name] = AgentNotification(
-                    agent=agent.name,
-                    timestamp=datetime.now(),
-                    level="info",
-                    summary="No activity recorded",
-                )
-
-        notifications = list(latest.values())
-
-        # Build custom format with lock counts
-        if not notifications:
-            return "━━━ No notifications ━━━"
-
-        lines = ["━━━ Agent Status ━━━"]
-        for n in notifications:
-            icon = notification_manager.STATUS_ICONS.get(n.level, "?")
-
-            # Get lock info for this agent
-            lock_info = ""
-            if lock_manager:
-                locks = lock_manager.get_locks_by_agent(n.agent)
-                lock_count = len(locks)
-                if lock_count == 0:
-                    lock_info = "[0 locks]"
-                elif lock_count == 1:
-                    lock_info = f"[1 lock: {locks[0][:12]}]"
-                else:
-                    lock_info = f"[{lock_count} locks]"
-
-            # Format: agent (12 chars) | icon | summary (truncated) | lock info
-            agent_name = n.agent[:12].ljust(12)
-            summary = n.summary[:20].ljust(20) if len(n.summary) > 20 else n.summary.ljust(20)
-            lines.append(f"{agent_name} {icon} {summary} {lock_info}")
-
-        lines.append("━━━━━━━━━━━━━━━━━━━━")
-        formatted = "\n".join(lines)
-
-        logger.info(f"Generated status summary for {len(notifications)} agents")
-        return formatted
-
-    except Exception as e:
-        logger.error(f"Error generating status summary: {e}")
-        return f"Error: {e}"
-
-
-@mcp.tool()
-async def notify(
-    agent: str,
-    level: str,
-    summary: str,
-    context: Optional[str] = None,
-    action_hint: Optional[str] = None,
-    ctx: Context = None,
-) -> str:
-    """Manually add a notification for an agent.
-
-    Use this to record significant events like task completion,
-    errors encountered, or when an agent needs attention.
-
-    Args:
-        agent: The agent name
-        level: One of: info, warning, error, success, blocked
-        summary: Brief one-line summary (max 100 chars)
-        context: Optional additional context
-        action_hint: Optional suggested next action
-    """
-    notification_manager = ctx.request_context.lifespan_context["notification_manager"]
-    logger = ctx.request_context.lifespan_context["logger"]
-
-    try:
-        await notification_manager.add_simple(
-            agent=agent,
-            level=level,
-            summary=summary,
-            context=context,
-            action_hint=action_hint,
-        )
-        logger.info(f"Added notification for {agent}: [{level}] {summary}")
-        return f"Notification added for {agent}"
-
-    except Exception as e:
-        logger.error(f"Error adding notification: {e}")
-        return f"Error: {e}"
 
 
 # ============================================================================
 # WAIT FOR AGENT TOOLS
+# See iterm_mcpy/tools/wait.py
 # ============================================================================
 
-@mcp.tool()
-async def wait_for_agent(request: WaitForAgentRequest, ctx: Context) -> str:
-    """Wait for an agent to complete or reach idle state.
-
-    This allows an orchestrator to wait for a subagent to finish its current
-    task. If the wait times out, returns a progress summary so you can decide
-    whether to wait longer or take action.
-
-    Args:
-        request: Contains agent name, timeout, and output options
-
-    Returns:
-        WaitResult with completion status, elapsed time, and optional output/summary
-    """
-    terminal = ctx.request_context.lifespan_context["terminal"]
-    agent_registry = ctx.request_context.lifespan_context["agent_registry"]
-    notification_manager = ctx.request_context.lifespan_context["notification_manager"]
-    logger = ctx.request_context.lifespan_context["logger"]
-
-    try:
-        req = ensure_model(WaitForAgentRequest, request)
-
-        # Find the agent
-        agent = agent_registry.get_agent(req.agent)
-        if not agent:
-            return WaitResult(
-                agent=req.agent,
-                completed=False,
-                timed_out=False,
-                elapsed_seconds=0,
-                status="unknown",
-                summary=f"Agent '{req.agent}' not found",
-                can_continue_waiting=False,
-            ).model_dump_json(indent=2, exclude_none=True)
-
-        # Get the session
-        session = await terminal.get_session_by_id(agent.session_id)
-        if not session:
-            return WaitResult(
-                agent=req.agent,
-                completed=False,
-                timed_out=False,
-                elapsed_seconds=0,
-                status="unknown",
-                summary=f"Session for agent '{req.agent}' not found",
-                can_continue_waiting=False,
-            ).model_dump_json(indent=2, exclude_none=True)
-
-        logger.info(f"Waiting up to {req.wait_up_to}s for agent {req.agent}")
-
-        # Capture initial output for comparison
-        initial_output = await session.get_screen_contents()
-
-        # Poll for completion
-        start_time = time.time()
-        poll_interval = 0.5  # Check every 500ms
-        last_output = initial_output
-
-        while True:
-            elapsed = time.time() - start_time
-
-            # Check if timed out
-            if elapsed >= req.wait_up_to:
-                # Timed out - generate summary
-                current_output = await session.get_screen_contents()
-
-                summary = None
-                if req.summary_on_timeout:
-                    # Generate a simple summary based on output changes
-                    if current_output != initial_output:
-                        lines = current_output.strip().split('\n')
-                        last_lines = lines[-3:] if len(lines) > 3 else lines
-                        summary = f"Still running. Last output: {' | '.join(last_lines)}"
-                    else:
-                        summary = "No output change detected during wait period"
-
-                # Add notification
-                await notification_manager.add_simple(
-                    agent=req.agent,
-                    level="info",
-                    summary=f"Wait timed out after {int(elapsed)}s",
-                    context=summary,
-                )
-
-                result = WaitResult(
-                    agent=req.agent,
-                    completed=False,
-                    timed_out=True,
-                    elapsed_seconds=elapsed,
-                    status="running",
-                    output=current_output if req.return_output else None,
-                    summary=summary,
-                    can_continue_waiting=True,
-                )
-                logger.info(f"Wait for {req.agent} timed out after {elapsed:.1f}s")
-                return result.model_dump_json(indent=2, exclude_none=True)
-
-            # Check if processing has stopped (idle)
-            is_processing = getattr(session, 'is_processing', False)
-            if not is_processing:
-                # Check if output has stabilized
-                current_output = await session.get_screen_contents()
-                if current_output == last_output:
-                    # Agent appears idle
-                    await notification_manager.add_simple(
-                        agent=req.agent,
-                        level="success",
-                        summary=f"Completed after {int(elapsed)}s",
-                    )
-
-                    result = WaitResult(
-                        agent=req.agent,
-                        completed=True,
-                        timed_out=False,
-                        elapsed_seconds=elapsed,
-                        status="idle",
-                        output=current_output if req.return_output else None,
-                        summary="Agent completed successfully",
-                        can_continue_waiting=False,
-                    )
-                    logger.info(f"Agent {req.agent} completed after {elapsed:.1f}s")
-                    return result.model_dump_json(indent=2, exclude_none=True)
-
-                last_output = current_output
-
-            await asyncio.sleep(poll_interval)
-
-    except Exception as e:
-        logger.error(f"Error waiting for agent: {e}")
-        return WaitResult(
-            agent=request.agent if hasattr(request, 'agent') else "unknown",
-            completed=False,
-            timed_out=False,
-            elapsed_seconds=0,
-            status="error",
-            summary=str(e),
-            can_continue_waiting=False,
-        ).model_dump_json(indent=2, exclude_none=True)
+# Re-export for backward compatibility with tests that import from this module.
+from iterm_mcpy.tools.wait import wait_for_agent  # noqa: E402,F401
 
 
 # ============================================================================
 # FEEDBACK SYSTEM TOOLS
+# See iterm_mcpy/tools/feedback.py
 # ============================================================================
-
-@mcp.tool()
-async def submit_feedback(
-    ctx: Context,
-    title: str,
-    description: str,
-    category: str = "enhancement",
-    agent_name: Optional[str] = None,
-    session_id: Optional[str] = None,
-    reproduction_steps: Optional[List[str]] = None,
-    suggested_improvement: Optional[str] = None,
-    error_messages: Optional[List[str]] = None,
-) -> str:
-    """Submit feedback about the iTerm MCP system.
-
-    This is the manual /feedback command. Use when you have suggestions,
-    found bugs, or want to request improvements to the iterm-mcp.
-
-    Args:
-        title: Short summary of the feedback
-        description: Detailed description of the issue or suggestion
-        category: One of: bug, enhancement, ux, performance, docs
-        agent_name: Name of the agent submitting (auto-detected if not provided)
-        session_id: Session ID (auto-detected from active session if not provided)
-        reproduction_steps: Steps to reproduce (for bugs)
-        suggested_improvement: What you think should be improved
-        error_messages: Any error messages encountered
-    """
-    feedback_registry = ctx.request_context.lifespan_context["feedback_registry"]
-    agent_registry = ctx.request_context.lifespan_context["agent_registry"]
-    notification_manager = ctx.request_context.lifespan_context["notification_manager"]
-    logger = ctx.request_context.lifespan_context["logger"]
-
-    try:
-        # Get agent info
-        if not agent_name and session_id:
-            agent = agent_registry.get_agent_by_session(session_id)
-            if agent:
-                agent_name = agent.name
-
-        if not session_id:
-            session_id = agent_registry.active_session or "unknown"
-
-        if not agent_name:
-            agent = agent_registry.get_agent_by_session(session_id)
-            agent_name = agent.name if agent else "unknown-agent"
-
-        # Collect context
-        collector = FeedbackCollector()
-        context = await collector.capture_context(
-            project_path=os.getcwd(),
-            recent_tool_calls=[],  # Would need hook integration for real data
-            recent_errors=error_messages or [],
-        )
-
-        # Parse category
-        try:
-            cat = FeedbackCategory(category.lower())
-        except ValueError:
-            cat = FeedbackCategory.ENHANCEMENT
-
-        # Create feedback entry
-        entry = FeedbackEntry(
-            agent_id=agent_name,
-            agent_name=agent_name,
-            session_id=session_id,
-            trigger_type=FeedbackTriggerType.MANUAL,
-            context=context,
-            category=cat,
-            title=title,
-            description=description,
-            reproduction_steps=reproduction_steps,
-            suggested_improvement=suggested_improvement,
-            error_messages=error_messages,
-        )
-
-        # Save to registry (sync method, no await needed)
-        feedback_registry.add(entry)
-
-        # Notify
-        await notification_manager.add_simple(
-            agent=agent_name,
-            level="success",
-            summary=f"Feedback submitted: {title[:50]}",
-            context=f"Feedback ID: {entry.id}",
-        )
-
-        logger.info(f"Feedback submitted: {entry.id} by {agent_name}")
-        return json.dumps({
-            "status": "submitted",
-            "feedback_id": entry.id,
-            "title": title,
-            "category": cat.value,
-            "message": "Thank you for your feedback! It has been recorded for review."
-        }, indent=2)
-
-    except Exception as e:
-        logger.error(f"Error submitting feedback: {e}")
-        return json.dumps({"error": str(e)}, indent=2)
-
-
-@mcp.tool()
-async def check_feedback_triggers(
-    ctx: Context,
-    agent_name: str,
-    session_id: str,
-    error_message: Optional[str] = None,
-    tool_call_name: Optional[str] = None,
-    output_text: Optional[str] = None,
-) -> str:
-    """Record events and check if feedback triggers should fire.
-
-    Call this to record errors, tool calls, or check for pattern matches
-    that might trigger feedback collection.
-
-    Args:
-        agent_name: Name of the agent
-        session_id: Session ID
-        error_message: Error message to record (triggers error threshold)
-        tool_call_name: Name of tool called (triggers periodic counter)
-        output_text: Text to scan for feedback patterns
-    """
-    hook_manager = ctx.request_context.lifespan_context["feedback_hook_manager"]
-    logger = ctx.request_context.lifespan_context["logger"]
-
-    try:
-        triggered = []
-        stats = hook_manager.get_stats(agent_name)
-
-        # Record error and check threshold - record_error returns trigger type if threshold reached
-        if error_message:
-            trigger_type = hook_manager.record_error(agent_name, error_message)
-            if trigger_type == FeedbackTriggerType.ERROR_THRESHOLD:
-                triggered.append({
-                    "trigger": "error_threshold",
-                    "reason": f"Error threshold reached ({stats['error_threshold']} errors)",
-                    "error": error_message,
-                })
-
-        # Record tool call and check periodic - record_tool_call returns trigger type if threshold reached
-        if tool_call_name:
-            trigger_type = hook_manager.record_tool_call(agent_name)
-            if trigger_type == FeedbackTriggerType.PERIODIC:
-                triggered.append({
-                    "trigger": "periodic",
-                    "reason": f"Periodic check ({stats['tool_call_threshold']} tool calls)",
-                })
-
-        # Check for pattern matches - check_pattern returns trigger type if pattern found
-        if output_text:
-            trigger_type = hook_manager.check_pattern(agent_name, output_text)
-            if trigger_type == FeedbackTriggerType.PATTERN_DETECTED:
-                triggered.append({
-                    "trigger": "pattern",
-                    "reason": "Feedback pattern detected in output",
-                })
-
-        logger.info(f"Trigger check for {agent_name}: {len(triggered)} triggers fired")
-        return json.dumps({
-            "agent": agent_name,
-            "triggers_fired": triggered,
-            "should_collect_feedback": len(triggered) > 0,
-        }, indent=2)
-
-    except Exception as e:
-        logger.error(f"Error checking triggers: {e}")
-        return json.dumps({"error": str(e)}, indent=2)
-
-
-@mcp.tool()
-async def query_feedback(
-    ctx: Context,
-    status: Optional[str] = None,
-    category: Optional[str] = None,
-    agent_id: Optional[str] = None,
-    limit: int = 20,
-) -> str:
-    """Query the feedback registry.
-
-    Args:
-        status: Filter by status (pending, triaged, in_progress, resolved, testing, closed)
-        category: Filter by category (bug, enhancement, ux, performance, docs)
-        agent_id: Filter by agent who submitted
-        limit: Max number of results
-    """
-    feedback_registry = ctx.request_context.lifespan_context["feedback_registry"]
-    logger = ctx.request_context.lifespan_context["logger"]
-
-    try:
-        # Parse filters
-        status_filter = None
-        if status:
-            try:
-                status_filter = FeedbackStatus(status.lower())
-            except ValueError:
-                pass
-
-        category_filter = None
-        if category:
-            try:
-                category_filter = FeedbackCategory(category.lower())
-            except ValueError:
-                pass
-
-        # Query
-        entries = await feedback_registry.query(
-            status=status_filter,
-            category=category_filter,
-            agent_id=agent_id,
-            limit=limit,
-        )
-
-        # Format results
-        results = []
-        for entry in entries:
-            results.append({
-                "id": entry.id,
-                "title": entry.title,
-                "category": entry.category.value,
-                "status": entry.status.value,
-                "agent": entry.agent_name,
-                "created_at": entry.created_at.isoformat(),
-                "github_issue_url": entry.github_issue_url,
-            })
-
-        logger.info(f"Query returned {len(results)} feedback entries")
-        return json.dumps({
-            "count": len(results),
-            "entries": results,
-        }, indent=2)
-
-    except Exception as e:
-        logger.error(f"Error querying feedback: {e}")
-        return json.dumps({"error": str(e)}, indent=2)
-
-
-@mcp.tool()
-async def fork_for_feedback(
-    ctx: Context,
-    feedback_id: str,
-    session_id: str,
-) -> str:
-    """Fork the current session to a git worktree for safe feedback submission.
-
-    Creates an isolated worktree and forks the Claude conversation there,
-    allowing the agent to provide detailed feedback without affecting
-    the main codebase.
-
-    Args:
-        feedback_id: The feedback ID to associate with the fork
-        session_id: The session ID to fork from
-    """
-    forker = ctx.request_context.lifespan_context["feedback_forker"]
-    notification_manager = ctx.request_context.lifespan_context["notification_manager"]
-    agent_registry = ctx.request_context.lifespan_context["agent_registry"]
-    logger = ctx.request_context.lifespan_context["logger"]
-
-    try:
-        agent = agent_registry.get_agent_by_session(session_id)
-        agent_name = agent.name if agent else "unknown"
-
-        # Create worktree
-        worktree_path = await forker.create_worktree(feedback_id)
-
-        # Get fork command (the actual forking is done by executing this command)
-        fork_command = forker.get_fork_command(session_id, worktree_path)
-
-        await notification_manager.add_simple(
-            agent=agent_name,
-            level="info",
-            summary=f"Forked for feedback: {feedback_id}",
-            context=f"Worktree: {worktree_path}",
-            action_hint="Continue in the forked session to provide feedback",
-        )
-
-        logger.info(f"Created worktree for session {session_id} at {worktree_path}")
-        return json.dumps({
-            "status": "worktree_created",
-            "feedback_id": feedback_id,
-            "worktree_path": str(worktree_path),
-            "fork_command": fork_command,
-            "message": "Worktree created. Execute the fork_command to continue in an isolated environment.",
-        }, indent=2)
-
-    except Exception as e:
-        logger.error(f"Error forking for feedback: {e}")
-        return json.dumps({"error": str(e)}, indent=2)
-
-
-@mcp.tool()
-async def triage_feedback_to_github(
-    ctx: Context,
-    feedback_id: str,
-    labels: Optional[List[str]] = None,
-    assignee: Optional[str] = None,
-) -> str:
-    """Create a GitHub issue from feedback.
-
-    Triages the feedback into a GitHub issue with proper labels and context.
-
-    Args:
-        feedback_id: The feedback ID to triage
-        labels: Additional labels for the issue
-        assignee: GitHub username to assign
-    """
-    feedback_registry = ctx.request_context.lifespan_context["feedback_registry"]
-    github_integration = ctx.request_context.lifespan_context["github_integration"]
-    notification_manager = ctx.request_context.lifespan_context["notification_manager"]
-    logger = ctx.request_context.lifespan_context["logger"]
-
-    try:
-        # Get the feedback entry (sync method, no await needed)
-        entry = feedback_registry.get(feedback_id)
-        if not entry:
-            return json.dumps({"error": f"Feedback {feedback_id} not found"}, indent=2)
-
-        # Create GitHub issue
-        issue_url = await github_integration.create_issue(
-            feedback=entry,
-            labels=labels,
-            assignee=assignee,
-        )
-
-        if issue_url:
-            # Update entry with issue URL (sync method, no await needed)
-            feedback_registry.update(
-                entry.id,
-                github_issue_url=issue_url,
-                status=FeedbackStatus.TRIAGED,
-            )
-
-            # Notify the agent
-            await notification_manager.add_simple(
-                agent=entry.agent_name,
-                level="success",
-                summary=f"Feedback triaged to GitHub",
-                context=issue_url,
-                action_hint="Check the GitHub issue for updates",
-            )
-
-            logger.info(f"Triaged feedback {feedback_id} to {issue_url}")
-            return json.dumps({
-                "status": "triaged",
-                "feedback_id": feedback_id,
-                "github_issue_url": issue_url,
-            }, indent=2)
-        else:
-            return json.dumps({
-                "status": "failed",
-                "error": "Failed to create GitHub issue. Check gh CLI is authenticated.",
-            }, indent=2)
-
-    except Exception as e:
-        logger.error(f"Error triaging feedback: {e}")
-        return json.dumps({"error": str(e)}, indent=2)
-
-
-@mcp.tool()
-async def notify_feedback_update(
-    ctx: Context,
-    feedback_id: str,
-    update_type: str,
-    message: str,
-    pr_url: Optional[str] = None,
-) -> str:
-    """Notify agents about feedback status updates.
-
-    Use this to notify the original agent when their feedback has been
-    addressed, a PR is ready for testing, etc.
-
-    Args:
-        feedback_id: The feedback ID
-        update_type: One of: acknowledged, in_progress, pr_opened, ready_for_testing, resolved
-        message: Human-readable update message
-        pr_url: URL to the PR if applicable
-    """
-    feedback_registry = ctx.request_context.lifespan_context["feedback_registry"]
-    notification_manager = ctx.request_context.lifespan_context["notification_manager"]
-    terminal = ctx.request_context.lifespan_context["terminal"]
-    agent_registry = ctx.request_context.lifespan_context["agent_registry"]
-    logger = ctx.request_context.lifespan_context["logger"]
-
-    try:
-        # Get the feedback entry (sync method, no await needed)
-        entry = feedback_registry.get(feedback_id)
-        if not entry:
-            return json.dumps({"error": f"Feedback {feedback_id} not found"}, indent=2)
-
-        # Update entry status based on update_type
-        status_map = {
-            "acknowledged": FeedbackStatus.TRIAGED,
-            "in_progress": FeedbackStatus.IN_PROGRESS,
-            "pr_opened": FeedbackStatus.IN_PROGRESS,
-            "ready_for_testing": FeedbackStatus.TESTING,
-            "resolved": FeedbackStatus.RESOLVED,
-        }
-
-        # Build updates dict
-        updates = {}
-        if update_type in status_map:
-            updates["status"] = status_map[update_type]
-
-        if pr_url:
-            updates["github_pr_url"] = pr_url
-
-        # Update entry (sync method, no await needed)
-        updated_entry = feedback_registry.update(entry.id, **updates)
-        if updated_entry:
-            entry = updated_entry
-
-        # Notify the agent
-        level = "success" if update_type == "ready_for_testing" else "info"
-        action_hint = None
-        if update_type == "ready_for_testing":
-            action_hint = f"Please test the fix: {pr_url}" if pr_url else "Please test the fix"
-
-        await notification_manager.add_simple(
-            agent=entry.agent_name,
-            level=level,
-            summary=f"Feedback update: {update_type}",
-            context=message,
-            action_hint=action_hint,
-        )
-
-        # Try to send a direct message to the agent's session if available
-        agent = agent_registry.get_agent(entry.agent_name)
-        if agent:
-            session = await terminal.get_session_by_id(agent.session_id)
-            if session:
-                # Don't execute, just display the notification
-                notification_text = f"\n[Feedback {feedback_id}] {update_type}: {message}"
-                if pr_url:
-                    notification_text += f"\nPR: {pr_url}"
-                # Log but don't send to terminal (could be disruptive)
-                logger.info(f"Would notify agent {entry.agent_name}: {notification_text}")
-
-        logger.info(f"Notified about feedback {feedback_id}: {update_type}")
-        return json.dumps({
-            "status": "notified",
-            "feedback_id": feedback_id,
-            "agent": entry.agent_name,
-            "update_type": update_type,
-            "new_status": entry.status.value,
-        }, indent=2)
-
-    except Exception as e:
-        logger.error(f"Error notifying about feedback: {e}")
-        return json.dumps({"error": str(e)}, indent=2)
-
-
-@mcp.tool()
-async def get_feedback_config(
-    ctx: Context,
-    update: bool = False,
-    error_threshold_count: Optional[int] = None,
-    periodic_tool_call_count: Optional[int] = None,
-    add_pattern: Optional[str] = None,
-    remove_pattern: Optional[str] = None,
-) -> str:
-    """Get or update feedback trigger configuration.
-
-    Args:
-        update: If True, apply the provided configuration changes
-        error_threshold_count: New error threshold (e.g., 3 = trigger after 3 errors)
-        periodic_tool_call_count: New periodic interval (e.g., 100 = trigger every 100 tool calls)
-        add_pattern: Regex pattern to add to pattern detection
-        remove_pattern: Regex pattern to remove from pattern detection
-    """
-    hook_manager = ctx.request_context.lifespan_context["feedback_hook_manager"]
-    logger = ctx.request_context.lifespan_context["logger"]
-
-    try:
-        if update:
-            # Apply updates
-            if error_threshold_count is not None:
-                hook_manager.config.error_threshold.count = error_threshold_count
-            if periodic_tool_call_count is not None:
-                hook_manager.config.periodic.tool_call_count = periodic_tool_call_count
-            if add_pattern:
-                hook_manager.config.pattern.patterns.append(add_pattern)
-            if remove_pattern and remove_pattern in hook_manager.config.pattern.patterns:
-                hook_manager.config.pattern.patterns.remove(remove_pattern)
-
-            # Save config
-            await hook_manager.save_config()
-            logger.info("Feedback config updated")
-
-        # Return current config
-        config = hook_manager.config
-        return json.dumps({
-            "enabled": config.enabled,
-            "error_threshold": {
-                "enabled": config.error_threshold.enabled,
-                "count": config.error_threshold.count,
-            },
-            "periodic": {
-                "enabled": config.periodic.enabled,
-                "tool_call_count": config.periodic.tool_call_count,
-            },
-            "pattern": {
-                "enabled": config.pattern.enabled,
-                "patterns": config.pattern.patterns,
-            },
-            "github": {
-                "repo": config.github.repo,
-                "default_labels": config.github.default_labels,
-            },
-        }, indent=2)
-
-    except Exception as e:
-        logger.error(f"Error with feedback config: {e}")
-        return json.dumps({"error": str(e)}, indent=2)
 
 
 # ============================================================================
@@ -3609,761 +2825,14 @@ async def get_feedback_config(
 
 # ============================================================================
 # MANAGER AGENT TOOLS
+# See iterm_mcpy/tools/managers.py
 # ============================================================================
 
-
-async def _execute_task_on_worker(
-    worker: str,
-    task: str,
-    timeout_seconds: Optional[int],
-    terminal: ItermTerminal,
-    agent_registry: AgentRegistry,
-    logger: logging.Logger,
-) -> tuple[Optional[str], bool, Optional[str]]:
-    """Execute a task on a worker agent and return result.
-
-    Args:
-        worker: Worker agent name
-        task: Command to execute
-        timeout_seconds: Optional timeout
-        terminal: Terminal instance
-        agent_registry: Agent registry
-        logger: Logger instance
-
-    Returns:
-        Tuple of (output, success, error)
-    """
-    agent = agent_registry.get_agent(worker)
-    if not agent:
-        return None, False, f"Worker agent '{worker}' not found"
-
-    session = await terminal.get_session_by_id(agent.session_id)
-    if not session:
-        return None, False, f"Session for worker '{worker}' not found"
-
-    try:
-        # Send the command
-        await session.send_text(task + "\n")
-
-        # Wait for command to complete with proper timeout
-        # Use a polling approach to check for command completion
-        wait_time = timeout_seconds if timeout_seconds else 30
-        poll_interval = 0.5
-        elapsed = 0.0
-
-        while elapsed < wait_time:
-            await asyncio.sleep(poll_interval)
-            elapsed += poll_interval
-            # Check if session is no longer processing (command completed)
-            if hasattr(session, 'is_processing') and not session.is_processing:
-                break
-
-        # Read output
-        output = await session.get_screen_contents(max_lines=100)
-
-        return output, True, None
-
-    except asyncio.TimeoutError:
-        return None, False, f"Task timed out after {timeout_seconds} seconds"
-    except Exception as e:
-        logger.error(f"Error executing task on worker {worker}: {e}")
-        return None, False, str(e)
-
-
-def _setup_manager_callbacks(
-    manager: ManagerAgent,
-    terminal: ItermTerminal,
-    agent_registry: AgentRegistry,
-    logger: logging.Logger,
-) -> None:
-    """Set up execution callbacks for a manager agent."""
-
-    async def execute_callback(
-        worker: str,
-        task: str,
-        timeout_seconds: Optional[int],
-    ) -> tuple[Optional[str], bool, Optional[str]]:
-        return await _execute_task_on_worker(
-            worker, task, timeout_seconds, terminal, agent_registry, logger
-        )
-
-    manager._execute_callback = execute_callback
-
-
-@mcp.tool()
-async def manage_managers(
-    request: ManageManagersRequest,
-    ctx: Context,
-) -> str:
-    """Manage manager agents with a single consolidated tool.
-
-    Consolidates: create_manager, list_managers, get_manager_info, remove_manager,
-                  add_worker_to_manager, remove_worker_from_manager
-
-    Operations:
-    - create: Create a new manager (requires manager_name)
-    - list: List all managers
-    - get_info: Get info about a manager (requires manager_name)
-    - remove: Remove a manager (requires manager_name)
-    - add_worker: Add a worker to a manager (requires manager_name, worker_name)
-    - remove_worker: Remove a worker from a manager (requires manager_name, worker_name)
-
-    Args:
-        request: The manager operation request with operation type and parameters
-
-    Returns:
-        JSON with operation results
-    """
-    terminal = ctx.request_context.lifespan_context["terminal"]
-    agent_registry = ctx.request_context.lifespan_context["agent_registry"]
-    manager_registry: ManagerRegistry = ctx.request_context.lifespan_context["manager_registry"]
-    logger = ctx.request_context.lifespan_context["logger"]
-
-    try:
-        if request.operation == "create":
-            if not request.manager_name:
-                response = ManageManagersResponse(
-                    operation=request.operation,
-                    success=False,
-                    error="manager_name is required for create operation"
-                )
-                return response.model_dump_json(indent=2, exclude_none=True)
-
-            # Convert worker roles from strings to SessionRole
-            worker_roles = {}
-            for worker, role_str in request.worker_roles.items():
-                worker_roles[worker] = ManagerSessionRole(role_str)
-
-            # Create the manager
-            manager = manager_registry.create_manager(
-                name=request.manager_name,
-                workers=request.workers,
-                delegation_strategy=DelegationStrategy(request.delegation_strategy),
-                worker_roles=worker_roles,
-                metadata=request.metadata,
-            )
-
-            # Set up execution callbacks
-            _setup_manager_callbacks(manager, terminal, agent_registry, logger)
-
-            logger.info(f"Created manager '{request.manager_name}' with {len(request.workers)} workers")
-
-            response = ManageManagersResponse(
-                operation=request.operation,
-                success=True,
-                data={
-                    "name": manager.name,
-                    "workers": manager.workers,
-                    "delegation_strategy": manager.strategy.value,
-                    "created": True
-                }
-            )
-            return response.model_dump_json(indent=2, exclude_none=True)
-
-        elif request.operation == "list":
-            managers = manager_registry.list_managers()
-
-            result = []
-            for manager in managers:
-                result.append({
-                    "name": manager.name,
-                    "workers": manager.workers,
-                    "delegation_strategy": manager.strategy.value,
-                    "worker_count": len(manager.workers),
-                })
-
-            logger.info(f"Listed {len(managers)} managers")
-            response = ManageManagersResponse(
-                operation=request.operation,
-                success=True,
-                data={"managers": result, "count": len(result)}
-            )
-            return response.model_dump_json(indent=2, exclude_none=True)
-
-        elif request.operation == "get_info":
-            if not request.manager_name:
-                response = ManageManagersResponse(
-                    operation=request.operation,
-                    success=False,
-                    error="manager_name is required for get_info operation"
-                )
-                return response.model_dump_json(indent=2, exclude_none=True)
-
-            manager = manager_registry.get_manager(request.manager_name)
-            if not manager:
-                response = ManageManagersResponse(
-                    operation=request.operation,
-                    success=False,
-                    error=f"Manager '{request.manager_name}' not found"
-                )
-                return response.model_dump_json(indent=2, exclude_none=True)
-
-            response = ManageManagersResponse(
-                operation=request.operation,
-                success=True,
-                data={
-                    "name": manager.name,
-                    "workers": manager.workers,
-                    "worker_roles": {k: v.value for k, v in manager.worker_roles.items()},
-                    "delegation_strategy": manager.strategy.value,
-                    "created_at": manager.created_at.isoformat(),
-                    "metadata": manager.metadata,
-                }
-            )
-            return response.model_dump_json(indent=2, exclude_none=True)
-
-        elif request.operation == "remove":
-            if not request.manager_name:
-                response = ManageManagersResponse(
-                    operation=request.operation,
-                    success=False,
-                    error="manager_name is required for remove operation"
-                )
-                return response.model_dump_json(indent=2, exclude_none=True)
-
-            removed = manager_registry.remove_manager(request.manager_name)
-
-            if removed:
-                logger.info(f"Removed manager '{request.manager_name}'")
-            else:
-                logger.warning(f"Manager '{request.manager_name}' not found")
-
-            response = ManageManagersResponse(
-                operation=request.operation,
-                success=removed,
-                data={"manager_name": request.manager_name}
-            )
-            return response.model_dump_json(indent=2, exclude_none=True)
-
-        elif request.operation == "add_worker":
-            if not request.manager_name or not request.worker_name:
-                response = ManageManagersResponse(
-                    operation=request.operation,
-                    success=False,
-                    error="manager_name and worker_name are required for add_worker operation"
-                )
-                return response.model_dump_json(indent=2, exclude_none=True)
-
-            manager = manager_registry.get_manager(request.manager_name)
-            if not manager:
-                response = ManageManagersResponse(
-                    operation=request.operation,
-                    success=False,
-                    error=f"Manager '{request.manager_name}' not found"
-                )
-                return response.model_dump_json(indent=2, exclude_none=True)
-
-            role = ManagerSessionRole(request.worker_role) if request.worker_role else None
-            manager.add_worker(request.worker_name, role)
-
-            logger.info(f"Added worker '{request.worker_name}' to manager '{request.manager_name}'")
-
-            response = ManageManagersResponse(
-                operation=request.operation,
-                success=True,
-                data={
-                    "manager_name": request.manager_name,
-                    "worker_name": request.worker_name,
-                    "role": request.worker_role
-                }
-            )
-            return response.model_dump_json(indent=2, exclude_none=True)
-
-        elif request.operation == "remove_worker":
-            if not request.manager_name or not request.worker_name:
-                response = ManageManagersResponse(
-                    operation=request.operation,
-                    success=False,
-                    error="manager_name and worker_name are required for remove_worker operation"
-                )
-                return response.model_dump_json(indent=2, exclude_none=True)
-
-            manager = manager_registry.get_manager(request.manager_name)
-            if not manager:
-                response = ManageManagersResponse(
-                    operation=request.operation,
-                    success=False,
-                    error=f"Manager '{request.manager_name}' not found"
-                )
-                return response.model_dump_json(indent=2, exclude_none=True)
-
-            removed = manager.remove_worker(request.worker_name)
-
-            if removed:
-                logger.info(f"Removed worker '{request.worker_name}' from manager '{request.manager_name}'")
-            else:
-                logger.warning(f"Worker '{request.worker_name}' not found in manager '{request.manager_name}'")
-
-            response = ManageManagersResponse(
-                operation=request.operation,
-                success=removed,
-                data={
-                    "manager_name": request.manager_name,
-                    "worker_name": request.worker_name
-                }
-            )
-            return response.model_dump_json(indent=2, exclude_none=True)
-
-        else:
-            response = ManageManagersResponse(
-                operation=request.operation,
-                success=False,
-                error=f"Unknown operation: {request.operation}"
-            )
-            return response.model_dump_json(indent=2, exclude_none=True)
-
-    except Exception as e:
-        logger.error(f"Error in manage_managers: {e}")
-        response = ManageManagersResponse(
-            operation=request.operation,
-            success=False,
-            error=str(e)
-        )
-        return response.model_dump_json(indent=2, exclude_none=True)
-
-
-@mcp.tool()
-async def delegate_task(
-    request: DelegateTaskRequest,
-    ctx: Context,
-) -> str:
-    """Delegate a task through a manager to an appropriate worker.
-
-    The manager selects a worker based on its delegation strategy and the
-    required role. The task is executed and optionally validated.
-
-    Args:
-        request: Task delegation request with manager, task, and options
-
-    Returns:
-        JSON with task execution result
-    """
-    terminal = ctx.request_context.lifespan_context["terminal"]
-    agent_registry = ctx.request_context.lifespan_context["agent_registry"]
-    manager_registry: ManagerRegistry = ctx.request_context.lifespan_context["manager_registry"]
-    logger = ctx.request_context.lifespan_context["logger"]
-
-    try:
-        manager = manager_registry.get_manager(request.manager)
-        if not manager:
-            return json.dumps({"error": f"Manager '{request.manager}' not found"}, indent=2)
-
-        # Ensure callbacks are set up
-        _setup_manager_callbacks(manager, terminal, agent_registry, logger)
-
-        # Convert role string to ManagerSessionRole if provided
-        role = ManagerSessionRole(request.role) if request.role else None
-
-        # Delegate the task
-        result = await manager.delegate(
-            task=request.task,
-            required_role=role,
-            validation=request.validation,
-            timeout_seconds=request.timeout_seconds,
-            retry_count=request.retry_count,
-        )
-
-        logger.info(f"Task delegated via manager '{request.manager}': {result.status.value}")
-
-        response = TaskResultResponse(
-            task_id=result.task_id,
-            task=result.task,
-            worker=result.worker,
-            status=result.status.value,
-            success=result.success,
-            output=result.output,
-            error=result.error,
-            duration_seconds=result.duration_seconds,
-            validation_passed=result.validation_passed,
-            validation_message=result.validation_message,
-        )
-        return response.model_dump_json(indent=2, exclude_none=True)
-
-    except Exception as e:
-        logger.error(f"Error delegating task: {e}")
-        return json.dumps({"error": str(e)}, indent=2)
-
-
-@mcp.tool()
-async def execute_plan(
-    request: ExecutePlanRequest,
-    ctx: Context,
-) -> str:
-    """Execute a multi-step task plan through a manager.
-
-    The manager orchestrates the execution of multiple steps, handling
-    dependencies and parallel execution as specified in the plan.
-
-    Args:
-        request: Plan execution request with manager and plan specification
-
-    Returns:
-        JSON with plan execution results
-    """
-    terminal = ctx.request_context.lifespan_context["terminal"]
-    agent_registry = ctx.request_context.lifespan_context["agent_registry"]
-    manager_registry: ManagerRegistry = ctx.request_context.lifespan_context["manager_registry"]
-    logger = ctx.request_context.lifespan_context["logger"]
-
-    try:
-        manager = manager_registry.get_manager(request.manager)
-        if not manager:
-            return json.dumps({"error": f"Manager '{request.manager}' not found"}, indent=2)
-
-        # Ensure callbacks are set up
-        _setup_manager_callbacks(manager, terminal, agent_registry, logger)
-
-        # Convert plan spec to TaskPlan
-        steps = []
-        for step_spec in request.plan.steps:
-            role = ManagerSessionRole(step_spec.role) if step_spec.role else None
-            step = TaskStep(
-                id=step_spec.id,
-                task=step_spec.task,
-                role=role,
-                optional=step_spec.optional,
-                depends_on=step_spec.depends_on,
-                validation=step_spec.validation,
-                timeout_seconds=step_spec.timeout_seconds,
-                retry_count=step_spec.retry_count,
-            )
-            steps.append(step)
-
-        plan = TaskPlan(
-            name=request.plan.name,
-            description=request.plan.description,
-            steps=steps,
-            parallel_groups=request.plan.parallel_groups,
-            stop_on_failure=request.plan.stop_on_failure,
-        )
-
-        # Execute the plan
-        plan_result = await manager.orchestrate(plan)
-
-        logger.info(
-            f"Plan '{plan.name}' completed: success={plan_result.success}, "
-            f"steps={len(plan_result.results)}"
-        )
-
-        # Convert results to response
-        result_responses = []
-        for result in plan_result.results:
-            result_responses.append(TaskResultResponse(
-                task_id=result.task_id,
-                task=result.task,
-                worker=result.worker,
-                status=result.status.value,
-                success=result.success,
-                output=result.output,
-                error=result.error,
-                duration_seconds=result.duration_seconds,
-                validation_passed=result.validation_passed,
-                validation_message=result.validation_message,
-            ))
-
-        response = PlanResultResponse(
-            plan_name=plan_result.plan_name,
-            success=plan_result.success,
-            results=result_responses,
-            duration_seconds=plan_result.duration_seconds,
-            stopped_early=plan_result.stopped_early,
-            stop_reason=plan_result.stop_reason,
-        )
-        return response.model_dump_json(indent=2, exclude_none=True)
-
-    except Exception as e:
-        logger.error(f"Error executing plan: {e}")
-        return json.dumps({"error": str(e)}, indent=2)
 
 # ============================================================================
 # ROLE MANAGEMENT TOOLS
+# See iterm_mcpy/tools/roles.py
 # ============================================================================
-
-
-@mcp.tool()
-async def assign_session_role(
-    ctx: Context,
-    session_id: str,
-    role: str,
-    assigned_by: Optional[str] = None,
-) -> str:
-    """Assign a role to a session for tool access control.
-
-    Args:
-        session_id: The iTerm session ID to assign the role to
-        role: The role name (devops, builder, debugger, researcher, tester, orchestrator, monitor, custom)
-        assigned_by: Optional agent name that is assigning this role (must have can_modify_roles permission)
-    """
-    role_manager: RoleManager = ctx.request_context.lifespan_context["role_manager"]
-    agent_registry: AgentRegistry = ctx.request_context.lifespan_context["agent_registry"]
-    logger = ctx.request_context.lifespan_context["logger"]
-
-    try:
-        # Permission check: if assigned_by is provided, verify they have can_modify_roles
-        if assigned_by:
-            caller_agent = agent_registry.get_agent(assigned_by)
-            if caller_agent:
-                caller_session_id = caller_agent.session_id
-                if not role_manager.can_modify_roles(caller_session_id):
-                    return json.dumps({
-                        "error": f"Agent '{assigned_by}' does not have permission to modify roles. "
-                                 "Only sessions with can_modify_roles=True (e.g., orchestrator) can assign roles."
-                    }, indent=2)
-
-        # Convert string to SessionRole enum
-        try:
-            session_role = SessionRole(role.lower())
-        except ValueError:
-            valid_roles = [r.value for r in SessionRole]
-            return json.dumps({
-                "error": f"Invalid role '{role}'. Valid roles are: {valid_roles}"
-            }, indent=2)
-
-        assignment = role_manager.assign_role(
-            session_id=session_id,
-            role=session_role,
-            assigned_by=assigned_by,
-        )
-
-        logger.info(f"Assigned role {role} to session {session_id}")
-
-        return json.dumps({
-            "status": "success",
-            "session_id": session_id,
-            "role": assignment.role.value,
-            "description": assignment.role_config.description,
-            "can_spawn_agents": assignment.role_config.can_spawn_agents,
-            "can_modify_roles": assignment.role_config.can_modify_roles,
-            "priority": assignment.role_config.priority,
-            "assigned_at": assignment.assigned_at.isoformat(),
-            "assigned_by": assignment.assigned_by,
-        }, indent=2)
-    except Exception as e:
-        logger.error(f"Error assigning role: {e}")
-        return json.dumps({"error": str(e)}, indent=2)
-
-@mcp.tool()
-async def get_session_role(
-    ctx: Context,
-    session_id: str,
-) -> str:
-    """Get the role assignment for a session.
-
-    Args:
-        session_id: The iTerm session ID to get the role for
-    """
-    role_manager: RoleManager = ctx.request_context.lifespan_context["role_manager"]
-    logger = ctx.request_context.lifespan_context["logger"]
-
-    try:
-        description = role_manager.describe(session_id)
-        return json.dumps(description, indent=2)
-    except Exception as e:
-        logger.error(f"Error getting session role: {e}")
-        return json.dumps({"error": str(e)}, indent=2)
-
-
-@mcp.tool()
-async def remove_session_role(
-    ctx: Context,
-    session_id: str,
-    removed_by: Optional[str] = None,
-) -> str:
-    """Remove the role assignment from a session.
-
-    Args:
-        session_id: The iTerm session ID to remove the role from
-        removed_by: Optional agent name that is removing this role (must have can_modify_roles permission)
-    """
-    role_manager: RoleManager = ctx.request_context.lifespan_context["role_manager"]
-    agent_registry: AgentRegistry = ctx.request_context.lifespan_context["agent_registry"]
-    logger = ctx.request_context.lifespan_context["logger"]
-
-    try:
-        # Permission check: if removed_by is provided, verify they have can_modify_roles
-        if removed_by:
-            caller_agent = agent_registry.get_agent(removed_by)
-            if caller_agent:
-                caller_session_id = caller_agent.session_id
-                if not role_manager.can_modify_roles(caller_session_id):
-                    return json.dumps({
-                        "error": f"Agent '{removed_by}' does not have permission to modify roles. "
-                                 "Only sessions with can_modify_roles=True (e.g., orchestrator) can remove roles."
-                    }, indent=2)
-
-        removed = role_manager.remove_role(session_id)
-        if removed:
-            logger.info(f"Removed role from session {session_id}")
-            return json.dumps({
-                "status": "success",
-                "session_id": session_id,
-                "message": "Role removed successfully"
-            }, indent=2)
-        else:
-            return json.dumps({
-                "status": "not_found",
-                "session_id": session_id,
-                "message": "No role was assigned to this session"
-            }, indent=2)
-    except Exception as e:
-        logger.error(f"Error removing session role: {e}")
-        return json.dumps({"error": str(e)}, indent=2)
-
-@mcp.tool()
-async def list_session_roles(
-    ctx: Context,
-    role_filter: Optional[str] = None,
-) -> str:
-    """List all session role assignments.
-
-    Args:
-        role_filter: Optional role name to filter by
-    """
-    role_manager: RoleManager = ctx.request_context.lifespan_context["role_manager"]
-    logger = ctx.request_context.lifespan_context["logger"]
-
-    try:
-        filter_role = None
-        if role_filter:
-            try:
-                filter_role = SessionRole(role_filter.lower())
-            except ValueError:
-                valid_roles = [r.value for r in SessionRole]
-                return json.dumps({
-                    "error": f"Invalid role filter '{role_filter}'. Valid roles are: {valid_roles}"
-                }, indent=2)
-
-        assignments = role_manager.list_roles(role_filter=filter_role)
-
-        result = []
-        for assignment in assignments:
-            result.append({
-                "session_id": assignment.session_id,
-                "role": assignment.role.value,
-                "description": assignment.role_config.description,
-                "priority": assignment.role_config.priority,
-                "assigned_at": assignment.assigned_at.isoformat(),
-                "assigned_by": assignment.assigned_by,
-            })
-
-        return json.dumps({
-            "count": len(result),
-            "assignments": result,
-        }, indent=2)
-    except Exception as e:
-        logger.error(f"Error listing session roles: {e}")
-        return json.dumps({"error": str(e)}, indent=2)
-
-@mcp.tool()
-async def list_available_roles(
-    ctx: Context,
-) -> str:
-    """List all available session roles and their default configurations."""
-    from core.models import DEFAULT_ROLE_CONFIGS
-
-    try:
-        roles = []
-        for role in SessionRole:
-            config = DEFAULT_ROLE_CONFIGS.get(role)
-            if config:
-                roles.append({
-                    "role": role.value,
-                    "description": config.description,
-                    "available_tools": config.available_tools,
-                    "restricted_tools": config.restricted_tools,
-                    "default_commands": config.default_commands,
-                    "can_spawn_agents": config.can_spawn_agents,
-                    "can_modify_roles": config.can_modify_roles,
-                    "priority": config.priority,
-                })
-            else:
-                roles.append({
-                    "role": role.value,
-                    "description": f"Custom role: {role.value}",
-                    "available_tools": [],
-                    "restricted_tools": [],
-                    "default_commands": [],
-                    "can_spawn_agents": False,
-                    "can_modify_roles": False,
-                    "priority": 3,
-                })
-
-        return json.dumps({
-            "count": len(roles),
-            "roles": roles,
-        }, indent=2)
-    except Exception as e:
-        return json.dumps({"error": str(e)}, indent=2)
-
-
-@mcp.tool()
-async def check_tool_permission(
-    ctx: Context,
-    session_id: str,
-    tool_name: str,
-) -> str:
-    """Check if a specific tool is allowed for a session based on its role.
-
-    Args:
-        session_id: The iTerm session ID to check
-        tool_name: The name of the tool to check permission for
-    """
-    role_manager: RoleManager = ctx.request_context.lifespan_context["role_manager"]
-    logger = ctx.request_context.lifespan_context["logger"]
-
-    try:
-        allowed, reason = role_manager.is_tool_allowed(session_id, tool_name)
-
-        assignment = role_manager.get_role(session_id)
-        role_info = {
-            "role": assignment.role.value if assignment else None,
-            "has_role": assignment is not None,
-        }
-
-        return json.dumps({
-            "session_id": session_id,
-            "tool_name": tool_name,
-            "allowed": allowed,
-            "reason": reason,
-            **role_info,
-        }, indent=2)
-    except Exception as e:
-        logger.error(f"Error checking tool permission: {e}")
-        return json.dumps({"error": str(e)}, indent=2)
-
-
-@mcp.tool()
-async def get_sessions_by_role(
-    ctx: Context,
-    role: str,
-) -> str:
-    """Get all session IDs that have a specific role assigned.
-
-    Args:
-        role: The role name to filter by
-    """
-    role_manager: RoleManager = ctx.request_context.lifespan_context["role_manager"]
-    logger = ctx.request_context.lifespan_context["logger"]
-
-    try:
-        try:
-            session_role = SessionRole(role.lower())
-        except ValueError:
-            valid_roles = [r.value for r in SessionRole]
-            return json.dumps({
-                "error": f"Invalid role '{role}'. Valid roles are: {valid_roles}"
-            }, indent=2)
-
-        session_ids = role_manager.get_sessions_by_role(session_role)
-
-        return json.dumps({
-            "role": session_role.value,
-            "count": len(session_ids),
-            "session_ids": session_ids,
-        }, indent=2)
-    except Exception as e:
-        logger.error(f"Error getting sessions by role: {e}")
-        return json.dumps({"error": str(e)}, indent=2)
 
 
 # ============================================================================
